requirements.txt
mysql-connector-python==8.0.25
requests==2.24.0
beautifulsoup4==4.9.3
selenium==3.141.0
python-dotenv==0.19.0
